Moved checkout state machine into the stepfunction folder
