+++ iam/policy-templates/dss-checkout-sfn-lambda.json
@@ -35,7 +35,9 @@
       ],
       "Resource": [
         "arn:aws:s3:::$DSS_S3_CHECKOUT_BUCKET",
-        "arn:aws:s3:::$DSS_S3_CHECKOUT_BUCKET/*"
+        "arn:aws:s3:::$DSS_S3_CHECKOUT_BUCKET/*",
+        "arn:aws:s3:::$DSS_S3_CHECKOUT_BUCKET_TEST",
+        "arn:aws:s3:::$DSS_S3_CHECKOUT_BUCKET_TEST/*"
       ]
     },
     {
